refactor(io): migrate datasets module Signed-off-by: Cameron Smith <cameron.ray.smith@gmail.com>
