+++ src/pyrovelocity/io/__init__.py
@@ -1,4 +1,8 @@
+import pyrovelocity.io.datasets
 from pyrovelocity.io.compressedpickle import CompressedPickle
 
 
-__all__ = ["CompressedPickle"]
+__all__ = [
+    "CompressedPickle",
+    "datasets",
+]
